+++ CHANGELOG.md
@@ -2,8 +2,9 @@
 FIX: Custom section names should now be decoded properly.
 IMPROVEMENT: bold and italic text now encoded as strong and em text (to match how Foundry VTT does it).
 FEATURE: Add option to include numbering on section headers, emulating the numbering in Realm Works.
-FEATURE: Add option for revealed topics, actors and tables to set their default permission to OBSERVER. (All snippets, except gm-directions, will be visible in revealed topics.)
 CHANGE: Connections and Content Links now use same topic name format as in Realm Works "topicname ( suffix - prefix )", rather than the format used in topic titles and governed content "prefix - topicname (suffix)".
+FEATURE: Add option to handle REVEALED state of things. If a topic is revealed, then its contents will be only the snippets that are revealed, and actors/tables/playlists will only be created from that topic if
+the containing snippets are revealed, and the topic will have its permissions set for players to be OBSERVERS. For topics which are not revealed, the full contents will be created (and the observer will NOT be set).
 
 1.3.1
 FIX: some relationships were not being decoded properly.